rename video_config files to videoConfig, to match our camelCase convention
