+++ packages/record/src/index.ts
@@ -2,7 +2,7 @@ import { VideoConsentPlugin } from "./consentVideo";
 import StartRecordPlugin from "./start";
 import StopRecordPlugin from "./stop";
 import TrialRecordExtension from "./trial";
-import VideoConfigPlugin from "./video_config";
+import VideoConfigPlugin from "./videoConfig";
 
 export default {
   TrialRecordExtension,
