+++ packages/record/src/videoConfig.ts
